requirements.txt
APScheduler==3.11.0
asyncpg==0.30.0
fastapi==0.115.8
fastapi-limiter==0.1.6
loguru==0.7.3
pydantic==2.10.6
pydantic-settings==2.7.1
PyJWT==1.7.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.38
uvicorn[standard]==0.34.0
alembic==1.14.1
duckdb==1.1.2
passlib==1.7.4
bcrypt==4.2.1
